Moved specs to right folder
