+++ superset/assets/spec/javascripts/explorev2/actions_spec.js
@@ -15,4 +15,15 @@ describe('reducers', () => {
       actions.setFieldValue('table', 'show_legend'));
     expect(newState.viz.form_data.show_legend).to.equal(false);
   });
+  it('adds a filter given a new filter', () => {
+    const newState = exploreReducer(initialState('table'),
+      actions.addFilter({
+        id: 1,
+        prefix: 'flt',
+        col: null,
+        op: null,
+        value: null,
+      }));
+    expect(newState.viz.form_data.filters).to.have.length(1);
+  });
 });
